Split up deployment
