+++ utils/deployment/NFTCollections.ts
@@ -11,8 +11,6 @@ delete NFTCollectionContracts["_BASE"];
 
 export async function deployNFTCollectionContract(name: string)
 {
-    const accounts = await ethers.getSigners();
-
     const NFTCollectionFactory = await ethers.getContractFactory(name);
     let NFTCollectionContract = await NFTCollectionFactory.deploy();
     await NFTCollectionContract.deployed();
