+++ utils/generators/addresses.json.ts
@@ -4,7 +4,7 @@ const addressesPath = "addresses.json";
 
 export const getAddressesJSON = () =>
 {
-    if (!fs.existsSync(addressesPath)) return {};
+    if (!fs.existsSync(addressesPath)) return { localhost : {} };
 
     const data = fs.readFileSync(addressesPath, { encoding : "utf8", flag : "r" });
     return JSON.parse(data.toString());
@@ -13,4 +13,11 @@ export const getAddressesJSON = () =>
 export const storeAddressesJSON = (obj : any) =>
 {
     fs.writeFileSync(addressesPath, JSON.stringify(obj));
+}
+
+export const clearAddressesJSON = () =>
+{
+    if (!fs.existsSync(addressesPath)) return;
+
+    fs.rmSync(addressesPath);
 }
