+++ utils/generators/graph-config.json.ts
@@ -1,0 +1,16 @@
+import fs from "fs";
+
+export const getGraphConfigJSONTemplate = (name : string, network : string = "localhost") =>
+{
+    return {
+        name : name
+    ,   output: "./graphs/{name}/generated/"
+    ,   chain: network
+    ,   datasources: []
+    }
+}
+
+export const storeGraphConfigJSON = (obj : { name : string }) =>
+{
+    fs.writeFileSync(`graphs/${obj.name}/config.json`, JSON.stringify(obj));
+}
